Added conversions for the metric/imperial system. Removed that sexy top bar that we all love so much.

+++ app/src/main/java/com/example/hadedahunter/ui/GlobalPreferences.kt
@@ -4,5 +4,5 @@ import androidx.lifecycle.ViewModel
 
 class GlobalPreferences : ViewModel() {
     var SelectedMeasuringSystem: String = ""
-    var MaximumDistance: Int = 0
+    var MaximumDistance: Double = 0.0
 }
